PyMaker/datas/object3/1st/i6401/Python_DOC_referPages.not-enough-values-to-unpack-expected-4-got-1-within-a-funtion-python.txt

Natural Text
I am new to python so I don't understand the Error I keep getting. I hope someone can explain. If you need any more information I'll edit it in.email,fname,sname,password=line.split(", ")#splits each line however creates an extra space between the lines because of enterValueError: not enough values to unpack (expected 4, got 1)I want it print like this:email@abc.org.uk Name Surname fakepasswordcheck@email.org.uk Z Y fakepasswordtesting@the.program Ray Check helloEdit: I've tried to remove the space between after the comma and tried to .split("\n") but gotemail,fname,sname,password=line.split("\n")#splits each line however creates an extra space between the lines because of enterValueError: not enough values to unpack (expected 4, got 2)At least I got one more value XD
Rather than re-implementing a csv reader, use the csv module from Python's standard library.Example copied from the documentation:This shows the canonical way to read a CSV file. Note that the csv.reader yields a list of strings for each iteration, so you don't have to split it.You can test the length of the line before splitting it:
The problem is your input. You need to make sure that each line in your input has all these 4 characteristics, and they are always delimited by a comma. This error occurs when the number of values on the right side on the equation do not match the number of variables on the right side. An example would be:The following works as the line is split into a list of 4 elementsBut the following does not workThere is no one to one mapping of elements to the variables. Hence, the interpreter will throw an error saying not enough values to unpack (expected 4, got 3). As 4 values were expected in list (as there are 4 variables on the left side of the equation, but only 3 were provided.


Answer URL
https://docs.python.org/3/library/csv.html
